Honor UtilizationType's `project_required` field
Currently, the UI limits setting utilization Project information to only
those utilizations with the "type" of "Consulting". Besides requiring
the hard-coding of the "Consulting" type's ID, this also incorrectly
limits the types of utilizations that may be assigned to specific
projects. The Nest API server already responds to requests for
"utilization types" with a boolean field designed for this purpose:
`project_required`.

Update the `UtilizationType` model to accept the `project_required`
field and update the UI to honor it.